making assemble command work
